+++ lib/src/app/app.dart
@@ -34,7 +34,7 @@ export 'appbuilder.dart';
 export 'http_client/app_http_client.dart';
 
 part 'app_credential.dart';
-part 'app_credential_service.dart';
+part 'app_credential_identity.dart';
 part 'app_dialog.dart';
 //part 'app_getxconnect.dart';
 
@@ -52,15 +52,13 @@ AppService get App => Get.find();
 
 class AppService extends GetxService {
   final AppInfo appInfo;
-  final AbstractCredentialService credential;
+  final AppCredential credential;
   final CredentialActions? _credentialActions;
   // final DialogSettings? _dialogSettings;
   // final SnackbarSettings _snackbarSettings;
-  final CredentialIdentity? _credentialIdentity;
   final AppDialog dialog;
 
-  late final CredentialIdentity identity =
-      _credentialIdentity ?? CredentialIdentity();
+  final AppCredentialIdentity identity;
 
   final AppSettings settings;
   //late final GetxConnect connect = GetxConnect._();
@@ -86,13 +84,12 @@ class AppService extends GetxService {
     required this.appInfo,
     required this.settings,
     required this.credential,
-    CredentialIdentity? credentialIdentity,
+    required this.identity,
     CredentialActions? credentialActions,
     required this.dialog,
     // DialogSettings? dialogSettings,
     // required SnackbarSettings notificationSettings,
-  })  : _credentialIdentity = credentialIdentity,
-        _credentialActions = credentialActions;
+  }) : _credentialActions = credentialActions;
   //   _dialogSettings = dialogSettings,
   //  _snackbarSettings = notificationSettings;
